Create function "operate" and global variables for operating.
Variables "number1", "number2", and "operator" are created as a future
storage for the user's inputs.

Function "operate" is created to work with the input variables and call
the diferent operation functions depending on the user's input.
The function stores the results of each operation within Array "result"
and returns the last element of it (the result from the latest operation).

The array "result" is created to work with it so the user can go back to
past results if he wishes.