Create function "execute" and modify function "operate" to work along it.
HTML and CSS files were corrected to reduce the calculator number of
buttons, matching better TOP's instructions.

General variables number1 and number2 were renamed to number1Arr and
number2Arr. Variables displayValue, resultsArr, and result are created
anew. Along with "operator", they are all now Array type variales, so
they can be updated from inside multiple-scope functions.

General constants "buttons", displayDiv", and "resultsDiv" are created
to connect to the respective DOM elements.

Function "operate" has been modified to work along main function
"execute". It now takes three arguments (number1, number2, and an
operator), and pushes the result into variales "resultsArr" and "result".

Function execute is created as the main function of the project.
It reacts to the user clicks and depending on the type of button (number,
operator, "=", "AC", "Erase" or "+/-" ), and reacts accordingly:

--> clicking a number will push it to number1Arr or number2Arr
depending on the conditions and display it on the lower div.

--> clicking an operator will push it to the operator array and display it
on the lower div. It will also execute "operate" whenever a number1 and
number2 are already created.

--> clicking "=" executes the operation whenever number1 and number2 exist.

--> clicking "AC" resets all variales and clears the display.

"+/-", ".", and "Erase" features are still pending to add to "execute".