Correct functionalities to display "x" and "," instead of "*" and ".".
Variables within function "execute" conditions that are used to display
characters are now created or updated using a ternary operator which
replaces "*" for "x" and "." for ",".

Global array "resultsArrToDisp" is created to store all the results but
the decimal ones will have a "," instead of a ".".

Function "operate" was corrected in order to work with these changes.
Inside it now variale solution is created to store the result of the
operations but replacing "." for "," in case they had decimals. Also,
all results are now cut to 2 decimals with the "toFixed" method.