Implement negativize and cap results' decimals to 2.
Implemented the negativize button functionality. Negativize button acts
as a number and a condition statement was added to the number cases
within "excecute" function so that when negativize button is pressed,
the current number the user is working with is add a "-" at the start.
If the number already is a negative, then negativize takes it out,
turning it into a positive (This method allows correction without having
to use the erase button and also avoids negative sign accumulation).

Re-worked the "operate" function by adding an if...else statement
inside with the number having decimals as a condition, in order to cap
the results to a maximum of 2 decimals. Also had to make it work with
substitution of "." for "," in the display.