Implement an error display when dividing by 0.
Modified division condition inside "operate" function so that when the
divisor is 0 it makes variable "solution" undefined and pushes it to
resultArrToDisp.

Modified all conditions that execute "operate" inside of "execute"
function, so that when resultArrToDisp.length-1 equals "undefined", an
error message is displayed explaining a division by 0 doesn't work,
clearing all variables to start anew in future operations.