Implement "erase" functionality
Implemented the "erase" functionality within function "execute".

Erase deletes de last element ("character") of number1Arr, operator, and
number2Arr arrays, and then displays it altogether so the last number is
deleted on screen. If the number present on screen is the result of the
previous operation, then it deletes the whole array and the whole number
on screen.