Fix use of "," button.
Previously, the "," button was treated as a number and that meant the
symbol could be used one after the other and also at the start of a
number.

A few conditionals have been added to the "number" conditions within
function "execute", so that only 1 "," symbol will be accepted per
number, and so that it won't be accepted as the starting character.